+++ src/main/java/org/monarchinitiative/lirical/likelihoodratio/CaseEvaluator.java
@@ -4,6 +4,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
+import org.monarchinitiative.lirical.analysis.Evolution;
 import org.monarchinitiative.lirical.analysis.Gene2Genotype;
 import org.monarchinitiative.lirical.exception.LiricalRuntimeException;
 import org.monarchinitiative.lirical.hpo.HpoCase;
@@ -23,9 +24,9 @@ import java.util.*;
 public class CaseEvaluator {
     private static final Logger logger = LoggerFactory.getLogger(CaseEvaluator.class);
     /** List of abnormalities seen in the person being evaluated. */
-    private final List<TermId> phenotypicAbnormalities;
+    private List<TermId> phenotypicAbnormalities;
     /** List of abnormalities excluded in the person being evaluated. */
-    private final List<TermId> negatedPhenotypicAbnormalities;
+    private List<TermId> negatedPhenotypicAbnormalities;
     /** Map of the observed genotypes in the VCF file. Key is an EntrezGene is, and the value is the average pathogenicity score times the
      * count of all variants in the pathogenic bin.*/
     private final Map<TermId,Gene2Genotype> genotypeMap;
@@ -53,6 +54,8 @@ public class CaseEvaluator {
     private final boolean useGenotypeAnalysis;
 
     private boolean verbose=true;
+
+    private List<LrWithExplanation> currentPhenotypeExplanation;
 
     private final static String EMPTY_STRING="";
 
@@ -138,20 +141,42 @@ public class CaseEvaluator {
     public void setVerbosity(boolean v) { this.verbose=v;}
 
 
-    private TestResult getResultNoDiseaseGene(HpoDisease disease,double pretest, List<Double> observed, List<Double> excluded) {
-        return new TestResult(observed,excluded,disease,pretest);
+
+    public HpoCase optimizeByEvolution() {
+        Evolution evolution = new Evolution(phenotypicAbnormalities,negatedPhenotypicAbnormalities);
+        int Nchroms = evolution.getNumberOfChromosomes();
+        while (evolution.nextGeneration()) {
+            for (int c=0;c<Nchroms;c++) {
+                phenotypicAbnormalities = evolution.getObservedTermsForChromosomeN(c);
+                negatedPhenotypicAbnormalities = evolution.getNegatedTermsForChromosomeN(c);
+                HpoCase newcase = evaluate();
+                double maxscore = newcase.getBestPosteriorProbability();
+                evolution.setScoreForChromosomeN(c,maxscore);
+            }
+            evolution.createNextGeneration();
+        }
+        phenotypicAbnormalities = evolution.getBestChromosomeObserved();
+        negatedPhenotypicAbnormalities = evolution.getBestChromosomeNegated();
+        return evaluate();
     }
 
-    private TestResult getResultDiseaseGeneNoVariantsFound(HpoDisease disease,double pretest, List<Double> observed, List<Double> excluded) {
-        return new TestResult(observed,excluded,disease,pretest);
-    }
+
+
 
 
     private List<Double> observedPhenotypesLikelihoodRatios(TermId diseaseId) {
         ImmutableList.Builder<Double> builderObserved = new ImmutableList.Builder<>();
+        HpoDisease disease = this.diseaseMap.get(diseaseId);
+        InducedDiseaseGraph idg = new InducedDiseaseGraph(disease,ontology);
         for (TermId tid : this.phenotypicAbnormalities) {
-            double LR = phenotypeLRevaluator.getLikelihoodRatio(tid, diseaseId);
-            builderObserved.add(LR);
+            LrWithExplanation lrwe = phenotypeLRevaluator.getLikelihoodRatio(tid, idg);
+            builderObserved.add(lrwe.getLR());
+            this.currentPhenotypeExplanation.add(lrwe);
+            /*logger.error("{}: {} {} [{}]",
+                    diseaseMap.get(diseaseId).getName(),
+                    ontology.getTermMap().get(tid).getName(),
+                    lrwe.getLR(),
+                    lrwe.getExplanation(ontology));*/
         }
         return builderObserved.build();
     }
@@ -173,6 +198,7 @@ public class CaseEvaluator {
     private  Map<TermId,TestResult>  phenotypeOnlyEvaluation() {
         ImmutableMap.Builder<TermId,TestResult> mapbuilder = new ImmutableMap.Builder<>();
         for (TermId diseaseId : diseaseMap.keySet()) {
+            this.currentPhenotypeExplanation=new ArrayList<>();
             HpoDisease disease = this.diseaseMap.get(diseaseId);
             double pretest = pretestProbabilityMap.get(diseaseId);
             List<Double> observedLR = observedPhenotypesLikelihoodRatios(diseaseId);
@@ -196,6 +222,8 @@ public class CaseEvaluator {
         List<Double> observedLR = observedPhenotypesLikelihoodRatios(diseaseId);
         List<Double> excludedLR = excludedPhenotypesLikelihoodRatios(diseaseId);
         TestResult result= new TestResult(observedLR, excludedLR, disease, pretest);
+        String phenoExp = getPhenotypeExplanation();
+        result.setPhenotypeExplanation(phenoExp);
         return Optional.of(result);
     }
 
@@ -260,14 +288,16 @@ public class CaseEvaluator {
             String expl = String.format("No variants found in disease-associated gene%s: %s",
                     genesWithNoIdentifiedVariant.size() > 1 ? "s" : "",
                     String.join("; ", genesWithNoIdentifiedVariant));
-            result.appendToExplanation(expl);
+            result.setGenotypeExplanation(expl);
 
         } else {
             // if we get here, then foundPredictedPathogenicVariant is true.
             Gene2Genotype g2g = this.genotypeMap.get(geneId);
             String exp = getGenotypeScoreExplanation(g2g, inheritancemodes,geneId);
-            result.appendToExplanation(exp);
+            result.setGenotypeExplanation(exp);
         }
+        String phenoExp = getPhenotypeExplanation();
+        result.setPhenotypeExplanation(phenoExp);
         return Optional.of(result);
     }
 
@@ -289,6 +319,18 @@ public class CaseEvaluator {
         return expl;
     }
 
+
+    private String getPhenotypeExplanation() {
+        ImmutableList.Builder<String> builder = new ImmutableList.Builder<>();
+        Collections.sort(this.currentPhenotypeExplanation);
+        for (LrWithExplanation lrwe:this.currentPhenotypeExplanation) {
+            String e = lrwe.getEscapedExplanation(this.ontology);
+            builder.add(e);
+        }
+        List<String> l = builder.build();
+        return String.join(": ",l);
+    }
+
     /**
      * Calculate the likelihood ratio for diseaseId. If there is no predicted pathogenic variant in the exome/genome file,
      * then we will return Optional.empty(), which will cause this diseases to be skipped in the differential diagnosis.
@@ -300,6 +342,7 @@ public class CaseEvaluator {
         double pretest = pretestProbabilityMap.get(diseaseId);
         List<Double> observedLR = observedPhenotypesLikelihoodRatios(diseaseId);
         List<Double> excludedLR = excludedPhenotypesLikelihoodRatios(diseaseId);
+        String phenoExp = getPhenotypeExplanation();
         TestResult result;
         Collection<TermId> associatedGenes = disease2geneMultimap.get(diseaseId);
         if (associatedGenes.isEmpty()) {
@@ -309,6 +352,7 @@ public class CaseEvaluator {
                 // keep differentials with no associated gene
                 // we create the TestResult based solely on the Phenotype data.
                 result = new TestResult(observedLR, excludedLR, disease, pretest);
+                result.setPhenotypeExplanation(phenoExp);
                 return Optional.of(result);
             } else {
                 // we skip this differential because there is no associated gene
@@ -361,7 +405,8 @@ public class CaseEvaluator {
             result = new TestResult(observedLR, excludedLR, disease, genotypeLR, geneId, pretest);
             Gene2Genotype g2g = this.genotypeMap.get(geneId);
             String exp  = getGenotypeScoreExplanation(g2g, inheritancemodes,geneId);
-            result.appendToExplanation(exp);
+            result.setGenotypeExplanation(exp);
+            result.setPhenotypeExplanation(phenoExp);
             return Optional.of(result);
         }
     }
@@ -379,6 +424,7 @@ public class CaseEvaluator {
         ImmutableMap.Builder<TermId,TestResult> mapbuilder = new ImmutableMap.Builder<>();
         for (TermId diseaseId : diseaseMap.keySet()) {
             Optional<TestResult> optionalTestResult;
+            this.currentPhenotypeExplanation=new ArrayList<>();
             if (useGenotypeAnalysis) {
                 if (keepIfNoCandidateVariant) {
                     optionalTestResult = evaluateDiseaseKeepingAllCandidates(diseaseId);
@@ -428,11 +474,11 @@ public class CaseEvaluator {
         for (TestResult res : results) {
             rank++;
             res.setRank(rank);
-            if (verbose && rank<11) {
-                TermId diseaseCurie = res.getDiseaseCurie();
-                String name = diseaseMap.get(diseaseCurie).getName();
-                System.err.println(String.format("Rank #%d: %s [%s]",rank,name,diseaseCurie.getValue()));
-            }
+//            if (verbose && rank<11) {
+//                TermId diseaseCurie = res.getDiseaseCurie();
+//                String name = diseaseMap.get(diseaseCurie).getName();
+//                //System.err.println(String.format("Rank #%d: %s [%s]",rank,name,diseaseCurie.getValue()));
+//            }
         }
         return resultMap;
     }